refactor: route name modify (#3)
