+++ apps/meetie-service/src/components/Study/CheckBox.tsx
