+++ apps/meetie-service/src/components/Study/Explorer/ExplorerTab.tsx
@@ -1,12 +1,12 @@
 "use client";
 import Image from "next/image";
 import { useState } from "react";
-import PageConter from "./PageCounter";
-import HashTag from "./HashTag";
-import CheckBox from "./CheckBox";
-import StudyCard from "./StudyCard";
-import MemberList from "./MemberList";
-import Filter from "./Filter";
+import PageConter from "../PageCounter";
+import HashTag from "../HashTag";
+import CheckBox from "../CheckBox";
+import StudyCard from "../StudyCard";
+import MemberList from "../Member/MemberList";
+import Filter from "../Filter";
 
 const Tab = () => {
   const [activeTab, setActiveTab] = useState("study");
