+++ apps/meetie-service/src/components/Study/StudyCard.tsx
