+++ apps/meetie-service/src/lib/contants.ts
@@ -5,25 +5,25 @@ export const GNB_DATA = [
   {
     text: "스터디룸",
     link: "/study-room",
-    icon: "svg/ic-gnb-study-room.svg",
+    icon: "/svg/ic-gnb-study-room.svg",
     disabledIcon: "svg/ic-gnb-study-room-disabled.svg",
   },
   {
     text: "탐색",
     link: "/search",
-    icon: "svg/ic-gnb-search.svg",
+    icon: "/svg/ic-gnb-search.svg",
     disabledIcon: "svg/ic-gnb-search-disabled.svg",
   },
   {
     text: "커뮤니티",
     link: "/community",
-    icon: "svg/ic-gnb-community.svg",
+    icon: "/svg/ic-gnb-community.svg",
     disabledIcon: "svg/ic-gnb-community-disabled.svg",
   },
   {
     text: "마이",
     link: "/my",
-    icon: "svg/ic-gnb-my.svg",
+    icon: "/svg/ic-gnb-my.svg",
     disabledIcon: "svg/ic-gnb-my-disabled.svg",
   },
 ];
